close #99: ResultSearch View, VC 분리 및 코드 리팩토링
