+++ Zatch/domain/ViewControllers/Search/ResultSearchViewController.swift
@@ -15,154 +15,29 @@ class ResultSearchViewController: BaseViewController, UIGestureRecognizerDelegat
     
     var resultData: [String] = []
     
-    //MARK: UIComponent
-    
-    class SearchCateogryDotButton: UIButton {
+    lazy var mainView = ResultSearchView().then{
+        $0.myCategoryButton.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(openCategoryBottomSheet)))
+        $0.myZatch.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(openMyZatchBottomSheet)))
+        $0.myZatch.addGestureRecognizer(UILongPressGestureRecognizer(target: self, action: #selector(textFieldDidPressedLong)))
         
-        //MARK: - LifeCycle
-        override init(frame: CGRect) {
-            super.init(frame: .zero)
-            
-            self.setImage(UIImage(named: "search_dot"), for: .normal)
-            self.setImage(UIImage(named: "search_dot_checked"), for: .selected)
-        }
+        $0.wantCategoryButton.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(openCategoryBottomSheet)))
+        $0.wantZatch.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(openWantZatchBottomSheet)))
+        $0.wantZatch.addGestureRecognizer(UILongPressGestureRecognizer(target: self, action: #selector(textFieldDidPressedLong)))
         
-        required init?(coder: NSCoder) {
-            fatalError("init(coder:) has not been implemented")
-        }
-        
+        $0.townFrame.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(townFrameDidClicked)))
     }
     
-    let topView = UIView()
-    
-    let separateLine = UIView().then{
-        $0.backgroundColor = .black10
-    }
-    
-    let separateRectangle = UIView().then{
-        $0.backgroundColor = .black5
-    }
-    
-    let myStackView = UIStackView().then{
-        $0.axis = .vertical
-        $0.alignment = .trailing
-        $0.spacing = 14
-    }
-    
-    lazy var myCategoryButton = SearchCateogryDotButton().then{
-        $0.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(openCategoryBottomSheet)))
-    }
-    
-    lazy var myZatch = UILabel().then{
-        $0.numberOfLines = 1
-        $0.textAlignment = .right
-        $0.textColor = .black85
-        $0.font = UIFont.pretendard(size: 16, family: .Bold)
-        $0.isUserInteractionEnabled = true
-        $0.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(openMyZatchBottomSheet)))
-        $0.addGestureRecognizer(UILongPressGestureRecognizer(target: self, action: #selector(textFieldDidPressedLong)))
-    }
-    
-    lazy var myZatchTextField = UITextField().then{
-        $0.isHidden = true
-        $0.textAlignment = .center
-        $0.textColor = .black85
-        $0.font = UIFont.pretendard(size: 16, family: .Bold)
-        $0.addPadding()
-        $0.tintColor = .black10
-        $0.returnKeyType = .done
-        
-        $0.layer.cornerRadius = 14
-        $0.backgroundColor = .white
-        $0.layer.shadowRadius = 5.0
-        $0.layer.shadowColor = UIColor(red: 0, green: 0, blue: 0, alpha: 0.1).cgColor
-        $0.layer.shadowOffset = CGSize(width: 0, height: 0)
-        $0.layer.shadowOpacity = 1
-        $0.layer.masksToBounds = false
-    }
-    
-    let wantStackView = UIStackView().then{
-        $0.axis = .vertical
-        $0.alignment = .leading
-        $0.spacing = 14
-    }
-    
-    lazy var wantCategoryButton = SearchCateogryDotButton().then{
-        $0.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(openCategoryBottomSheet)))
-    }
-    
-    lazy var wantZatch = UILabel().then{
-        $0.numberOfLines = 1
-        $0.textColor = .black85
-        $0.font = UIFont.pretendard(size: 16, family: .Bold)
-        $0.isUserInteractionEnabled = true
-        $0.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(openWantZatchBottomSheet)))
-        $0.addGestureRecognizer(UILongPressGestureRecognizer(target: self, action: #selector(textFieldDidPressedLong)))
-    }
-    
-    lazy var wantZatchTextField = UITextField().then{
-        $0.isHidden = true
-        $0.textAlignment = .center
-        $0.textColor = .black85
-        $0.font = UIFont.pretendard(size: 16, family: .Bold)
-        $0.addPadding()
-        $0.tintColor = .black10 //cursor 색상
-        $0.returnKeyType = .done
-        
-        $0.layer.cornerRadius = 14
-        $0.backgroundColor = .white
-        $0.layer.shadowRadius = 5.0
-        $0.layer.shadowColor = UIColor(red: 0, green: 0, blue: 0, alpha: 0.1).cgColor
-        $0.layer.shadowOffset = CGSize(width: 0, height: 0)
-        $0.layer.shadowOpacity = 1
-        $0.layer.masksToBounds = false
-    }
-    
-    let exchangeImage = UIImageView().then{
-        $0.image = UIImage(named: "exchange_vertical")
-    }
-    
-    lazy var townFrame = UIStackView().then{
-        $0.spacing = 4
-        $0.axis = .horizontal
-        $0.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(townFrameDidClicked)))
-    }
-    let townSelectedLabel = UILabel().then{
-        $0.text = "홍제동"
-        $0.font = UIFont.pretendard(size: 14, family: .Medium)
-        $0.textColor = .black85
-    }
-    let townSelectArrow = UIImageView().then{
-        $0.image = UIImage(named: "arrow_down")
-    }
-    
-    let searchFrame = UIStackView().then{
-        $0.spacing = 4
-        $0.axis = .horizontal
-    }
-    let searchFilterLabel = UILabel().then{
-        $0.text = "검색 필터"
-        $0.font = UIFont.pretendard(size: 14, family: .Medium)
-        $0.textColor = .black85
-    }
-    let searchFilterImage = UIImageView().then{
-        $0.image = UIImage(named: "search_filter")
-    }
-    
-    lazy var tableView = UITableView().then{
-        $0.separatorStyle = .none
-        
-        $0.register(ResultTableViewCell.self, forCellReuseIdentifier: "ResultTableViewCell")
-    }
-    
-    let emptyResultView = ResultEmptyTableViewCell()
     
     override func viewDidLoad() {
         
         super.viewDidLoad()
         
-        setUpView()
-        setUpConstraint()
+        self.view.addSubview(mainView)
+        
+        mainView.snp.makeConstraints{
+            $0.top.leading.trailing.bottom.equalToSuperview()
+        }
+    
         setUpDelegate()
         
         
@@ -170,7 +45,7 @@ class ResultSearchViewController: BaseViewController, UIGestureRecognizerDelegat
 //        if(resultData.isEmpty){
 //            emptyResultView.isHidden = false
 //        }else{
-            emptyResultView.isHidden = true
+        mainView.emptyResultView.isHidden = true
 //        }
          
     }
@@ -178,11 +53,11 @@ class ResultSearchViewController: BaseViewController, UIGestureRecognizerDelegat
     //TextField 입력 끝나거나 취소됐을 경우
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         
-        myZatchTextField.isHidden = true
-        wantZatchTextField.isHidden = true
+        mainView.myZatchTextField.isHidden = true
+        mainView.wantZatchTextField.isHidden = true
         
-        myZatch.isHidden = false
-        wantZatch.isHidden = false
+        mainView.myZatch.isHidden = false
+        mainView.wantZatch.isHidden = false
         
         self.view.endEditing(true)
     }
@@ -190,11 +65,11 @@ class ResultSearchViewController: BaseViewController, UIGestureRecognizerDelegat
     //MARK: - Helper
     func setUpDelegate(){
         
-        tableView.dataSource = self
-        tableView.delegate = self
+        mainView.tableView.dataSource = self
+        mainView.tableView.delegate = self
 
-        myZatchTextField.delegate = self
-        wantZatchTextField.delegate = self
+        mainView.myZatchTextField.delegate = self
+        mainView.wantZatchTextField.delegate = self
         
     }
     
@@ -206,7 +81,7 @@ class ResultSearchViewController: BaseViewController, UIGestureRecognizerDelegat
         bottomSheet.currentTag = myZatchIndex
         
         bottomSheet.selectCompleteHandelr = { text, index in
-            self.myZatch.text = text
+            self.mainView.myZatch.text = text
             self.myZatchIndex = index
         }
         
@@ -220,7 +95,7 @@ class ResultSearchViewController: BaseViewController, UIGestureRecognizerDelegat
         bottomSheet.currentTag = wantZatchIndex
         
         bottomSheet.selectCompleteHandelr = { text, index in
-            self.wantZatch.text = text
+            self.mainView.wantZatch.text = text
             self.wantZatchIndex = index
         }
         bottomSheet.loadViewIfNeeded()
@@ -234,7 +109,7 @@ class ResultSearchViewController: BaseViewController, UIGestureRecognizerDelegat
         
         vc.categorySelectHandler = { category in
             print(category)
-            (recognizer.view as? SearchCateogryDotButton)?.isSelected = true
+            (recognizer.view as? ResultSearchView.SearchCateogryDotButton)?.isSelected = true
         }
         
         vc.loadViewIfNeeded()
@@ -245,26 +120,26 @@ class ResultSearchViewController: BaseViewController, UIGestureRecognizerDelegat
     func textFieldDidPressedLong(_ recognizer : UIGestureRecognizer){
         
         //기존 입력값 초기화
-        myZatchTextField.text = nil
-        wantZatchTextField.text = nil
+        mainView.myZatchTextField.text = nil
+        mainView.wantZatchTextField.text = nil
         
-        myZatchTextField.placeholder = myZatch.text
-        wantZatchTextField.placeholder = wantZatch.text
+        mainView.myZatchTextField.placeholder = mainView.myZatch.text
+        mainView.wantZatchTextField.placeholder = mainView.wantZatch.text
         
-        myZatchTextField.setPlaceholderColor(.black10)
-        wantZatchTextField.setPlaceholderColor(.black10)
+        mainView.myZatchTextField.setPlaceholderColor(.black10)
+        mainView.wantZatchTextField.setPlaceholderColor(.black10)
         
-        myZatchTextField.isHidden = false
-        wantZatchTextField.isHidden = false
+        mainView.myZatchTextField.isHidden = false
+        mainView.wantZatchTextField.isHidden = false
         
-        myZatch.isHidden = true
-        wantZatch.isHidden = true
+        mainView.myZatch.isHidden = true
+        mainView.wantZatch.isHidden = true
         
         //커서 올리기
-        if(recognizer.view == wantZatch){
-            wantZatchTextField.becomeFirstResponder()
+        if(recognizer.view == mainView.wantZatch){
+            mainView.wantZatchTextField.becomeFirstResponder()
         }else{
-            myZatchTextField.becomeFirstResponder()
+            mainView.myZatchTextField.becomeFirstResponder()
         }
     }
     
@@ -306,16 +181,16 @@ extension ResultSearchViewController: UITextFieldDelegate{
     //return 키 클릭시 호출 메서드
     func textFieldShouldReturn(_ textField: UITextField) -> Bool {
         
-        myZatchTextField.isHidden = true
-        wantZatchTextField.isHidden = true
+        mainView.myZatchTextField.isHidden = true
+        mainView.wantZatchTextField.isHidden = true
         
-        myZatch.isHidden = false
-        wantZatch.isHidden = false
+        mainView.myZatch.isHidden = false
+        mainView.wantZatch.isHidden = false
         
-        if(textField == myZatchTextField){
-            myZatch.text = myZatchTextField.text
+        if(textField == mainView.myZatchTextField){
+            mainView.myZatch.text = mainView.myZatchTextField.text
         }else{
-            wantZatch.text = wantZatchTextField.text
+            mainView.wantZatch.text = mainView.wantZatchTextField.text
         }
         
         self.view.endEditing(true)
